Conflicts:
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/bookerstatements/client/BookerStatementsPresenter.java
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/formdata/client/search/FormSearchView.java
	mdb/src/main/java/com/aplana/sbrf/taxaccounting/mdb/RateMDB.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/refbook/impl/RefBookDepartment.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/refbook/impl/RefBookUniversal.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/DeclarationDataServiceImpl.java
	src/main/resources/form_template/vat/vat_724_2_2/v2014/content.xml
	src/main/resources/form_template/vat/vat_724_7/v2014/content.xml

+++ service/src/main/java/com/aplana/sbrf/taxaccounting/refbook/impl/RefBookUniversal.java
@@ -482,6 +482,7 @@ public class RefBookUniversal implements RefBookDataProvider {
                 }
 
                 boolean isJustNeedValuesUpdate = (versionFrom == null && versionTo == null);
+
                 //Получаем идентификатор записи справочника без учета версий
                 Long recordId = refBookDao.getRecordId(uniqueRecordId);
                 //Получаем еще неотредактированную версию
@@ -553,7 +554,7 @@ public class RefBookUniversal implements RefBookDataProvider {
                 //Обновление периода актуальности
                 if (isRelevancePeriodChanged) {
                     List<Long> uniqueIdAsList = Arrays.asList(uniqueRecordId);
-                    if (previousVersion.isVersionEndFake() && SimpleDateUtils.addDayToDate(previousVersion.getVersionEnd(), 1).equals(versionFrom)) {
+                    if (previousVersion != null && (previousVersion.isVersionEndFake() && SimpleDateUtils.addDayToDate(previousVersion.getVersionEnd(), 1).equals(versionFrom))) {
                         //Если установлена дата окончания, которая совпадает с существующей фиктивной версией - то она удаляется
                         Long previousVersionEnd = refBookDao.findRecord(refBookId, recordId, versionFrom);
                         refBookDao.deleteRecordVersions(REF_BOOK_RECORD_TABLE_NAME, Arrays.asList(previousVersionEnd));
